Move all TC*Info classes into TCInfo.h
